# Optimize SplFixedArray::toArray()